Add Basic Lobby Code
I added the basic functionality to create a lobby when you host a game. When you host a game, you will automatically go to the game scene so I will have to refactor this so that you stay in the same scene and ready up before you join the game scene together.